fix pr

+++ mType/environment/registry/builtin/HashCodeFunction.cpp
@@ -16,12 +16,12 @@ namespace environment::registry::builtin
             if constexpr (std::is_same_v<std::decay_t<decltype(value)>, std::string>)
             {
                 std::hash<std::string> hasher;
-                return static_cast<int>(hasher(value) & 0x7FFFFFFF);
+                return static_cast<int64_t>(hasher(value) & 0x7FFFFFFF);
             }
             else if constexpr (std::is_same_v<std::decay_t<decltype(value)>, value::InternedString>)
             {
                 std::hash<std::string> hasher;
-                return static_cast<int>(hasher(value.getString()) & 0x7FFFFFFF);
+                return static_cast<int64_t>(hasher(value.getString()) & 0x7FFFFFFF);
             }
             else if constexpr (std::is_same_v<std::decay_t<decltype(value)>, int64_t>)
             {
@@ -32,7 +32,7 @@ namespace environment::registry::builtin
             else if constexpr (std::is_same_v<std::decay_t<decltype(value)>, double>)
             {
                 std::hash<double> hasher;
-                return static_cast<int>(hasher(value) & 0x7FFFFFFF);
+                return static_cast<int64_t>(hasher(value) & 0x7FFFFFFF);
             }
             else if constexpr (std::is_same_v<std::decay_t<decltype(value)>, bool>)
             {
@@ -46,7 +46,7 @@ namespace environment::registry::builtin
                 }
                 std::string contentHash = value->getContentHash();
                 std::hash<std::string> hasher;
-                return static_cast<int>(hasher(contentHash) & 0x7FFFFFFF);
+                return static_cast<int64_t>(hasher(contentHash) & 0x7FFFFFFF);
             }
             else if constexpr (std::is_same_v<std::decay_t<decltype(value)>, nullptr_t>)
             {
